pyproject.toml
[project]
name = "openboard"
version = "0.1.0"
description = "An accessible and cross-platform chess GUI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accessible-output3",
    "blinker>=1.9.0",
    "chess>=1.11.2",
    "pydantic>=2.12.4",
    "wxpython>=4.2.4",
]

[project.scripts]
openboard = "openboard.views.views:main"


[dependency-groups]
dev = [
    "cairosvg>=2.7.0",
    "pillow>=10.0.0",
    "pyinstaller>=6.0.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
    "ty>=0.0.11",
]

[tool.uv.sources]
accessible-output3 = { git = "https://github.com/SigmaNight/accessible_output3" }
